refactor: minimal `near-crypto-hash`
