+++ near-crypto-hash/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "near-crypto-hash"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+categories.workspace = true
+repository.workspace = true
+homepage.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+bs58 = "0.5"
+
+serde = { version = "1", features = ["derive"], optional = true }
+borsh = { version = "1", features = ["derive"], optional = true }
+arbitrary = { version = "1.4", features = ["derive"], optional = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+schemars = { version = "0.8", optional = true }
+
+[features]
+default = []
+
+serde = ["dep:serde"]
+
+borsh = ["dep:borsh"]
+
+abi = ["borsh/unstable__schema", "dep:schemars"]
+
+arbitrary = ["dep:arbitrary"]
+
+near-contracts = ["serde", "borsh"]
+
+__near-sdk-unit-testing = ["near-contracts"]
