+++ near-global-contracts/Cargo.toml
@@ -10,13 +10,11 @@ rust-version.workspace = true
 
 [dependencies]
 near-account-id = "2.6.0"
-near-sdk-core = { path = "../near-sdk-core/", version = "0.1.0" }
+near-crypto-hash = { path = "../near-crypto-hash/", version = "~0.1.0" }
 
-sha3 = "0.10"
-bs58 = "0.5"
-base64 = "0.22"
-hex = "0.4"
+hex = { version = "0.4", optional = true }
 
+cfg_eval = { version = "0.1" }
 serde = { version = "1", features = ["derive"], optional = true }
 serde_json = { version = "1", features = ["preserve_order"], optional = true }
 serde_with = { version = "3", features = [
@@ -25,53 +23,59 @@ serde_with = { version = "3", features = [
     "json",
 ], optional = true }
 borsh = { version = "1.0.0", features = ["derive"], optional = true }
-schemars = { version = "0.8.22", optional = true }
 
 arbitrary = { version = "1.4", optional = true }
 
+sha3 = { version = "0.10", optional = true }
 near-env = { path = "../near-env/", version = "~0.1.0", optional = true }
 
 near-primitives-core = { version = "0.35", optional = true }
 
-[dev-dependencies]
-# XXX: `near-sdk` was added in order to enable tests and doctests compiling with mockchain
-near-sdk = { path = "../near-sdk", features = ["unit-testing"] }
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+schemars = { version = "0.8.22", optional = true }
 
 [features]
+default = []
+
 serde = [
     "dep:serde",
     "dep:serde_json",
     "dep:serde_with",
     "near-account-id/serde",
-    "near-sdk-core/serde",
+    "near-crypto-hash/serde",
 ]
 
-borsh = ["dep:borsh", "near-account-id/borsh", "near-sdk-core/borsh"]
+borsh = ["dep:borsh", "near-account-id/borsh", "near-crypto-hash/borsh"]
 
 abi = [
     "dep:schemars",
     "borsh",
     "borsh/unstable__schema",
     "near-account-id/abi",
-    "near-sdk-core/abi",
+    "near-crypto-hash/abi",
 ]
 
 arbitrary = [
     "dep:arbitrary",
-    "near-sdk-core/arbitrary",
     "near-account-id/arbitrary",
+    "near-crypto-hash/arbitrary",
 ]
 
-near-primitives-interop = [
-    "dep:near-primitives-core",
-    "near-sdk-core/near-primitives-interop",
-]
+near-primitives-interop = ["dep:near-primitives-core"]
 
-near-contracts = ["dep:near-env", "serde", "borsh"]
+digest = ["borsh", "dep:sha3", "dep:hex"]
+
+near-contracts = [
+    "dep:near-env",
+    "dep:hex",
+    "serde",
+    "borsh",
+    "near-crypto-hash/near-contracts",
+]
 
 __near-sdk-unit-testing = [
     "near-contracts",
     "near-primitives-interop",
-    "near-sdk-core/__near-sdk-unit-testing",
+    "near-crypto-hash/__near-sdk-unit-testing",
     "near-env/__near-sdk-unit-testing",
 ]
