+++ near-global-contracts/src/lib.rs
@@ -1,9 +1,10 @@
+#[cfg(all(feature = "near-contracts", feature = "digest"))]
+compile_error!(
+    "features `near-contracts` and `digest` are mutually exclusive - both provide keccak256 for `derive-account-id`"
+);
+
 mod global_contract_identifier;
 pub use global_contract_identifier::*;
 
 mod state_init;
 pub use state_init::*;
-
-#[cfg(test)]
-// XXX: `near-sdk` was added in order to enable tests and doctests compiling with mockchain
-use near_sdk as _;
