+++ near-sdk-core/Cargo.toml
@@ -32,6 +32,7 @@ arbitrary = { version = "1.4", optional = true }
 near-account-id = { version = "2.3.0" }
 near-gas = { version = "0.3" }
 near-token = { version = "0.3" }
+near-crypto-hash = { path = "../near-crypto-hash/", version = "~0.1.0" }
 near-env = { path = "../near-env/", version = "~0.1.0", optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
@@ -58,34 +59,48 @@ serde = [
     "near-account-id/serde",
     "near-gas/serde",
     "near-token/serde",
+    "near-crypto-hash/serde",
 ]
 borsh = [
     "dep:borsh",
     "near-account-id/borsh",
     "near-gas/borsh",
     "near-token/borsh",
+    "near-crypto-hash/borsh",
 ]
 
 abi = [
     "borsh/unstable__schema",
     "near-account-id/abi",
     "near-gas/abi",
+    "near-crypto-hash/abi",
     "serde_with/schemars_0_8",
     "dep:schemars",
 ]
 
-arbitrary = ["dep:arbitrary", "arbitrary/derive", "near-account-id/arbitrary"]
+arbitrary = [
+    "dep:arbitrary",
+    "arbitrary/derive",
+    "near-account-id/arbitrary",
+    "near-crypto-hash/arbitrary",
+]
 
 near-primitives-interop = ["dep:near-primitives-core"]
 near-crypto-interop = ["dep:near-crypto", "dep:unicode-segmentation"]
 
-near-contracts = ["serde", "borsh", "dep:near-env"]
+near-contracts = [
+    "serde",
+    "borsh",
+    "near-crypto-hash/near-contracts",
+    "dep:near-env",
+]
 
 __near-sdk-unit-testing = [
     "near-contracts",
     "near-primitives-interop",
     "near-crypto-interop",
     "near-env/__near-sdk-unit-testing",
+    "near-crypto-hash/__near-sdk-unit-testing",
     "dep:unicode-segmentation",
     "dep:near-parameters",
 ]
